Изменены типы данных

+++ plane.h
@@ -12,28 +12,28 @@ class WARSHIPS_AIRPLANES_DLLSHARED_EXPORT Plane : public QObject, public QGraphi
 {
     Q_OBJECT
 public:
-    Plane(qint32 id, PlayerInfo::Type type, qreal angle, qreal speed, qreal angleSpeed, qint16 health);
-    void setSpeed(qreal speed);
-    void setAngleSpeed(qreal angleSpeed);
-    void setHealth(qint16 health);
+    Plane(qint16 id, PlayerInfo::Type type, qint16 angle, quint8 speed, qint8 angleSpeed, quint8 health);
+    void setSpeed(quint8 speed);
+    void setAngleSpeed(qint8 angleSpeed);
+    void setHealth(quint8 health);
 
     PlayerInfo::Type    getType() const;
-    qreal   getSpeed() const;
-    qreal   getAngleSpeed() const;
-    qint16  getHealth() const;
-    qint32  getId() const;
+    quint8   getSpeed() const;
+    qint8    getAngleSpeed() const;
+    quint8   getHealth() const;
+    qint16   getId() const;
 
     virtual void advance(int phase);
     virtual QRectF boundingRect() const;
     virtual void paint(QPainter *painter, const QStyleOptionGraphicsItem *option, QWidget *widget);
 
 private:
-    qint32 id;
+    qint16 id;
     PlayerInfo::Type type;
-    qreal  speed;
-    qreal  angleSpeed;
-    qint16 health;
-    qreal  angle;
+    quint8  speed;
+    qint8   angleSpeed;
+    quint8  health;
+    qint16  angle;
 
 signals:
     void planeMoved(Plane* plane);
